+++ NEWS.md
@@ -1,9 +1,10 @@
 # mvtboost 0.3
 - Improved documentation for \code{mvtb}, and other functions
 - \code{predict} now drops unecessary dimensions
-- \code{cluster.covex} obtains a 'plot' argument, a call to \code{heat.covex}
-- \code{heat.covex} has better plotting defaults to allow room for y-axis labels. Gets a 'mar' argument, with a sensible default
-- \code{plot.perspec} plots variable names by default, gets \code{ylab, xlab, zlab} arguments
+- \code{cluster.covex} becomes \code{mvtb.cluster}, a more generic way to cluster any table. obtains a 'plot' argument, a call to \code{mvtb.heat}
+- \code{heat.covex} becomes \code{mvtb.heat}, a generic way to heatmap any table (e.g. covex or relative influence). Has better plotting defaults to allow room for y-axis labels. Gets 'mar','cexRow','cexCol' arguments, with sensible defaults. Clustering can also be turned off by setting clust.method=NULL.
+- \code{mvtb.perspec} plots variable names by default, gets \code{ylab, xlab, zlab} arguments
+- fixed weighting influence bug in \code{mvtb}
 
 
 # mvtboost 0.2.2